# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Hypatia Backend API
  description: >
    Contract between the Flutter frontend and the FastAPI backend for the
    Hypatia LLM wiki application. Covers health, Classes CRUD, file
    upload/serve/delete, wiki tree/page CRUD/search/export/lint/rebuild,
    chat WebSocket + history, task status, and class backup/import.
  version: "0.4.0"
servers:
  - url: http://127.0.0.1:8000
    description: >-
      Local dev backend. Port may vary 8000-8010 if 8000 is taken
      (see Task 1.9 auto-launch/port management).
paths:
...
  /api/classes:
    get:
      summary: List all Classes
      operationId: listClasses
      responses:
        "200":
          description: All Classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassRead"
    post:
      summary: Create a new Class
      operationId: createClass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassCreate"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassRead"
  /api/classes/{class_id}:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    get:
      summary: Get a single Class
      operationId: getClass
      responses:
        "200":
          description: Class details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassRead"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update a Class's name/description
      operationId: updateClass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdate"
      responses:
        "200":
          description: Updated Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassRead"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a Class and all its data
      operationId: deleteClass
      responses:
        "204":
          description: Class deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/classes/{class_id}/chat:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    get:
      summary: >-
        WebSocket chat connection (OpenAPI 3.0 has no native WebSocket
        support). See ChatWs* schemas for message formats.
      operationId: chatWebSocket
      responses:
        "101":
          description: Switching Protocols (WebSocket upgrade)
  /api/classes/{class_id}/chat/history:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    get:
      summary: Get paginated chat history
      operationId: getChatHistory
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Chat messages (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessageRead"
    delete:
      summary: Clear chat history for a Class
      operationId: clearChatHistory
      responses:
        "204":
          description: Chat history cleared
...
components:
  parameters:
    ClassId:
      name: class_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
      required: [detail, code]
    ClassCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required: [name]
    ClassUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    ClassRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, description, created_at, updated_at]
...
    ChatRole:
      type: string
      enum: [user, assistant, system]
    ChatMessageRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        class_id:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/ChatRole"
        content:
          type: string
        command:
          type: string
          nullable: true
        metadata_json:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, class_id, role, content, created_at, updated_at]

    # --- Chat WebSocket protocol -------------------------------------------
...